Use multithreaded mode for flask.
Multi-threaded mode should be used here, even for this very simple app.
Otherwise, if someone tests this container on e.g. AWS ECS or AWS EB
with an ELB, they will see the ELB health check keep-alive request
consume the only request thread, which will cause the container to be
killed repeatedly and the user to be deeply confused.

I lost many hours trying to debug this, figuring it was a problem with
how I had ECS/EB set up, but it was actually a nit in this example
flask app.